Adding additional install help
Added in the baserewrite statement to the htaccess files that is needed when opendcim isn't installed on the root and mod-rewrite is being a dick.  Also added a helper into the index.php to account for people that don't read the install instructions or are a developer that forget how it works when they look at someone else's system that is having issues.  Finally added in helper text to the preflight that talks about the rewrite base directive in the .htaccess

+++ index.php
@@ -1,4 +1,9 @@
 <?php
+	# catch for assholes that don't read the install instructions
+	if(!file_exists("db.inc.php")){
+		require_once( "preflight.inc.php" );
+		exit;
+	}
 /*	if ( ! $_SERVER["HTTPS"] ) {
 		printf( "<meta http-equiv='refresh' content='0; url='https://%s'>", $_SERVER["SERVER_NAME"] );
 		exit();
